Edited drawChart(), edited D3 chart selector
Edited the drawChart() function to be called with two parameters,
stressArr and stressId. The D3 selection of the .chart-container
occurs via the unique Id found on .js-single-result. This
ensures the chart will be rendered in the appropriate <div>. It
prevents the chart from being rendered in only the first <div>
with the .chart-container class.