CI testing triggered by push to any branch except master
CI testing is triggered by pushing to any branch except master.
The CI workflow will not run when there's a git push to master,
but will run for a git push to any other branch.

Adding new features and CI testing will done on a features branch.
If all the tests pass on the features branch, push to the master branch
to deploy. A different workflow will handle the deploy from master.